### 2021-02-10 #3026 @whedon
[PRE REVIEW]: starry_process: Interpretable Gaussian processes for stellar light curves
**Submitting author:** @rodluger (<a href="http://orcid.org/0000-0002-0296-3826">Rodrigo Luger</a>)
**Repository:** <a href="https://github.com/rodluger/starry_process" target ="_blank">https://github.com/rodluger/starry_process</a>
**Version:** v0.9.6
