+++ joss/bib.bib
@@ -1,570 +1,3 @@
-@article{Agol2020,
-  author        = {{Agol}, Eric and {Luger}, Rodrigo and {Foreman-Mackey}, Daniel},
-  title         = {{Analytic Planetary Transit Light Curves and Derivatives for Stars with Polynomial Limb Darkening}},
-  journal       = {\aj},
-  keywords      = {486, 489, 1709, 293, 1855, 38, 1711, 918, 442, 922, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics},
-  year          = 2020,
-  month         = mar,
-  volume        = {159},
-  number        = {3},
-  eid           = {123},
-  pages         = {123},
-  doi           = {10.3847/1538-3881/ab4fee},
-  archiveprefix = {arXiv},
-  eprint        = {1908.03222},
-  primaryclass  = {astro-ph.EP},
-  adsurl        = {https://ui.adsabs.harvard.edu/abs/2020AJ....159..123A},
-  adsnote       = {Provided by the SAO/NASA Astrophysics Data System}
-}
-
-@article{AlvarezCollado1989,
-  title   = {Rotation of real spherical harmonics},
-  journal = {Computer Physics Communications},
-  volume  = {52},
-  number  = {3},
-  pages   = {323 - 331},
-  year    = {1989},
-  issn    = {0010-4655},
-  doi     = {https://doi.org/10.1016/0010-4655(89)90107-0},
-  author  = {Jos\'e Ram\'on \'Alvarez Collado and Jaime Fern\'andez Rico and Rafael L\'opez and Miguel Paniagua and Guillermo Ram\'irez}
-}
-
-@article{Ambikasaran2015,
-  author        = {{Ambikasaran}, Sivaram and {Foreman-Mackey}, Daniel and {Greengard}, Leslie and {Hogg}, David W. and {O'Neil}, Michael},
-  title         = {{Fast Direct Methods for Gaussian Processes}},
-  journal       = {IEEE Transactions on Pattern Analysis and Machine Intelligence},
-  keywords      = {Mathematics - Numerical Analysis, Astrophysics - Instrumentation and Methods for Astrophysics, Mathematics - Statistics Theory, Mathematics - Numerical Analysis, Astrophysics - Instrumentation and Methods for Astrophysics, Mathematics - Statistics Theory},
-  year          = 2015,
-  month         = jun,
-  volume        = {38},
-  pages         = {252},
-  doi           = {10.1109/TPAMI.2015.2448083},
-  archiveprefix = {arXiv},
-  eprint        = {1403.6015},
-  primaryclass  = {math.NA},
-  adsurl        = {https://ui.adsabs.harvard.edu/abs/2015ITPAM..38..252A},
-  adsnote       = {Provided by the SAO/NASA Astrophysics Data System}
-}
-
-@article{Angus2018,
-  author        = {{Angus}, Ruth and {Morton}, Timothy and {Aigrain}, Suzanne and {Foreman-Mackey}, Daniel and {Rajpaul}, Vinesh},
-  title         = {{Inferring probabilistic stellar rotation periods using Gaussian processes}},
-  journal       = {\mnras},
-  keywords      = {methods: data analysis, methods: statistical, techniques: photometric, stars: rotation, stars: solar-type, starspots, Astrophysics - Solar and Stellar Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics},
-  year          = 2018,
-  month         = feb,
-  volume        = {474},
-  number        = {2},
-  pages         = {2094-2108},
-  doi           = {10.1093/mnras/stx2109},
[…]
