### 2021-02-16 d7e232f
Adding MCMC interface
