### 2021-02-16 3acc091
Update exceptions.py
More generous tolerance to fix bounds errors in the latitude parameters.
