### 2021-02-16 6670e0d
run examples only on change or release
