### 2021-02-16 95b44cd
fix examples triggers
