+++ .github/workflows/examples.yml
@@ -6,10 +6,8 @@ on:
       - published
   push:
     paths:
-      include:
-        - 'docs/notebooks/*.ipynb'
-      exclude:
-        - 'docs/notebooks/*.dev.ipynb'
+      - 'docs/notebooks/*.ipynb'
+      - '!docs/notebooks/*.dev.ipynb'
 
 jobs:
   notebooks:
