### 2021-02-16 d36e150
better initial state
