+++ docs/api.rst
@@ -5,6 +5,10 @@ API
     :members:
     :inherited-members:
 
+.. autoclass:: starry_process.MCMCInterface
+    :members:
+    :inherited-members:
+
 .. autofunction:: starry_process.beta2gauss
 
 .. autofunction:: starry_process.gauss2beta
