+++ joss/paper.md
@@ -1,0 +1,205 @@
+---
+title: "starry_process: Interpretable Gaussian processes for stellar light curves"
+tags:
+  - Python
+  - astronomy
+authors:
+  - name: Rodrigo Luger
+    orcid: 0000-0002-0296-3826
+    affiliation: "1, 2"
+  - name: Daniel Foreman-Mackey
+    orcid: 0000-0002-9328-5652
+    affiliation: 1
+  - name: Christina Hedges
+    orcid: 0000-0002-3385-8391
+    affiliation: "3, 4"
+affiliations:
+  - name: Center for Computational Astrophysics, Flatiron Institute, New York, NY
+    index: 1
+  - name: Virtual Planetary Laboratory, University of Washington, Seattle, WA
+    index: 2
+  - name: Bay Area Environmental Research Institute, P.O. Box 25, Moffett Field, CA 94035, USA
+    index: 3
+  - name: NASA Ames Research Center, Moffett Field, CA
+    index: 4
+date: 21 January 2021
+bibliography: bib.bib
+---
+
+#
+
+# Summary
+
+The `starry_process` code implements an interpretable Gaussian
+process (GP) for modeling variability in stellar light curves.
+As dark starspots rotate in and out of view, the total flux received
+from a distant star will change over time. Unresolved flux time series
+therefore encode information about the spatial
+structure of features on the stellar surface. The `starry_process` software
+package allows one to easily model the flux variability due to starspots,
+whether one is interested in understanding the properties of
+these spots or marginalizing over the stellar variability when it
+is treated as a nuisance signal.
+The main difference between the GP implemented here and typical
+GPs used to model stellar variability is the explicit dependence of
+our GP on physical properties of the star, such as its period, inclination,
+and limb darkening coefficients, and on properties of the spots, such as their
+radius and latitude distributions (see Figure \ref{fig:samples}).
+This code is the `Python` implementation of the interpretable
+GP algorithm developed in @PaperII.
+
+![Five random samples from our GP (columns) conditioned on two different
+hyperparameter vectors $\pmb{\theta}_\bullet$ (rows). The samples are
+shown on the surface of the star in a Mollweide projection alongside
+the corresponding light curves viewed at several different inclinations.
+(a) Samples from a GP describing a star with small mid-latitude spots. (b)
+Samples from a GP describing a star with larger high-latitude spots.
+\label{fig:samples}](figures/samples.pdf)
+
+#
+
+# Statement of need
+
+Mapping the surfaces of stars using time series measurements is a fundamental
+problem in modern time-domain stellar astrophysics. This inverse problem is
+ill-posed and computationally intractable, but in the associated AAS Journals
+publication submitted in parallel to this paper [@PaperII], we derive an interpretable
+effective Gaussian Process (GP) model for this problem that enables robust
+probabilistic characterization of stellar surfaces using photometric time series
+observations.
+Our model builds on previous work by @Perger2020 on semi-interpretable
+Gaussian processes for stellar timeseries data and by @Morris2020b
+on approximate inference for large ensembles of stellar light curves.
+Implementation of our model requires the efficient evaluation of
+a set of special functions and recursion relations that are not readily
+available in existing probabilistic programming frameworks. The `starry_process`
+package provides the necessary elements to perform this analysis with existing
+and forthcoming astronomical datasets.
+
+# Implementation
+
+We implement our interpretable GP in the user-friendly `Python`
+package `starry_process`, which can be installed via `pip` or from source on
+[GitHub](https://github.com/rodluger/starry_process). The code is thoroughly
+[unit-tested](https://github.com/rodluger/starry_process/tree/master/tests)
+and [well documented](https://starry_process.readthedocs.io), with
[…]
