### 2021-02-17 #3026 @arfon
:+1: thanks @rodluger. I suspect these folks might be pretty busy but let's try 🤞

:wave: @nespinoza @bmorris3 @jradavenport - would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
