### 2021-02-17 056a4a5
new notebook, fixed examples
