+++ docs/notebooks/notebook_config.py
@@ -1,0 +1,100 @@
+import matplotlib
+import matplotlib.pyplot as plt
+import numpy as np
+from corner import corner as _corner
+import warnings
+from IPython.display import set_matplotlib_formats
+from IPython import get_ipython
+
+# Inline hi-res plots
+get_ipython().run_line_magic("matplotlib", "inline")
+set_matplotlib_formats("retina")
+
+# Disable annoying font warnings
+matplotlib.font_manager._log.setLevel(50)
+
+# Style
+plt.style.use("default")
+plt.rcParams["savefig.dpi"] = 100
+plt.rcParams["figure.dpi"] = 100
+plt.rcParams["figure.figsize"] = (12, 4)
+plt.rcParams["font.size"] = 14
+plt.rcParams["text.usetex"] = False
+plt.rcParams["font.family"] = "sans-serif"
+plt.rcParams["font.sans-serif"] = ["Liberation Sans"]
+plt.rcParams["font.cursive"] = ["Liberation Sans"]
+try:
+    plt.rcParams["mathtext.fallback"] = "cm"
+except KeyError:
+    plt.rcParams["mathtext.fallback_to_cm"] = True
+plt.rcParams["mathtext.fallback_to_cm"] = True
+
+# Short arrays when printing
+np.set_printoptions(threshold=0)
+
+# Disable theano deprecation warnings
+warnings.filterwarnings("ignore", category=DeprecationWarning)
+warnings.filterwarnings(
+    "ignore", category=matplotlib.MatplotlibDeprecationWarning
+)
+warnings.filterwarnings("ignore", category=FutureWarning)
+warnings.filterwarnings("ignore", category=UserWarning, module="theano")
+
+
+# Override `corner.py` with some tweaks
+def corner(*args, **kwargs):
+    """
+    Override `corner.corner` by making some appearance tweaks.
+
+    """
+    # Get the usual corner plot
+    figure = _corner(*args, **kwargs)
+
+    # Get the axes
+    ndim = int(np.sqrt(len(figure.axes)))
+    axes = np.array(figure.axes).reshape((ndim, ndim))
+
+    # Smaller tick labels
+    for ax in axes[1:, 0]:
+        for tick in ax.yaxis.get_major_ticks():
+            tick.label.set_fontsize(10)
+        formatter = matplotlib.ticker.ScalarFormatter(useOffset=False)
+        ax.yaxis.set_major_formatter(formatter)
+        ax.set_ylabel(
+            ax.get_ylabel(), fontsize=kwargs.get("corner_label_size", 16)
+        )
+    for ax in axes[-1, :]:
+        for tick in ax.xaxis.get_major_ticks():
+            tick.label.set_fontsize(10)
+        formatter = matplotlib.ticker.ScalarFormatter(useOffset=False)
+        ax.xaxis.set_major_formatter(formatter)
+        ax.set_xlabel(
+            ax.get_xlabel(), fontsize=kwargs.get("corner_label_size", 16)
+        )
+
+    # Pad the axes to always include the truths
+    truths = kwargs.get("truths", None)
+    if truths is not None:
+        for row in range(1, ndim):
+            for col in range(row):
+                lo, hi = np.array(axes[row, col].get_xlim())
+                if truths[col] < lo:
+                    lo = truths[col] - 0.1 * (hi - truths[col])
+                    axes[row, col].set_xlim(lo, hi)
+                    axes[col, col].set_xlim(lo, hi)
+                elif truths[col] > hi:
+                    hi = truths[col] - 0.1 * (hi - truths[col])
+                    axes[row, col].set_xlim(lo, hi)
+                    axes[col, col].set_xlim(lo, hi)
+
+                lo, hi = np.array(axes[row, col].get_ylim())
[…]
