+++ .github/workflows/examples.yml
@@ -1,13 +1,10 @@
 name: examples
 
 on:
-  release:
-    types:
-      - published
   push:
-    paths:
-      - 'docs/notebooks/*.ipynb'
-      - '!docs/notebooks/*.dev.ipynb'
+    branches: [master]
+  pull_request:
+    branches: [master]
 
 jobs:
   notebooks:
@@ -25,6 +22,13 @@ jobs:
         with:
           fetch-depth: 0
           submodules: true
+
+      - name: Set up environment
+        shell: bash -l {0}
+        run: |
+          jupyter nbextension enable --py widgetsnbextension
+          sudo apt-get update
+          sudo apt-get install ffmpeg
 
       - uses: actions/cache@v2
         with:
@@ -55,11 +59,28 @@ jobs:
           python -m pip install -e ".[docs]"
           python -c 'import theano; print(theano.__version__)'
 
+      - name: Cache notebooks
+        id: cache-notebooks
+        uses: actions/cache@v2
+        with:
+          path: docs/notebooks
+          key: ${{ runner.os }}-${{ matrix.pymc-version }}-${{ github.sha }}
+          restore-keys: |
+            ${{ runner.os }}-${{ matrix.pymc-version }}-
+
       - name: Execute notebooks
         shell: bash -l {0}
         run: |
           cd docs/notebooks
-          for f in *.ipynb; do jupyter nbconvert --to notebook --execute $f --output $f; done
+          RUN_NOTEBOOKS=$([[ $(git log -1 --pretty=format:"%s") == *"[run-notebooks]"* ]] && echo 1 || echo 0)
+          echo "RUN_NOTEBOOKS:" $RUN_NOTEBOOKS
+          NOTEBOOKS_CHANGED=$([[ $(git diff --name-only HEAD^ HEAD) == *notebooks* ]] && echo 1 || echo 0)
+          echo "NOTEBOOKS_CHANGED:" $NOTEBOOKS_CHANGED
+          NEW_TAG=$([[ $GITHUB_REF == "refs/tags/v"* ]] && echo 1 || echo 0)
+          echo "NEW_TAG:" $NEW_TAG
+          if [[ $NOTEBOOKS_CHANGED == 1 || $RUN_NOTEBOOKS == 1 || NEW_TAG == 1 ]]; then
+            for f in *.ipynb; do jupyter nbconvert --to notebook --execute $f --output $f; done
+          fi
 
       - uses: actions/upload-artifact@v2
         with:
