+++ setup.cfg
@@ -3,4 +3,4 @@ setup_requires =
     setuptools_scm
 
 [metadata]
-license_files = LICENSE.txt
+license_files = LICENSE
