### 2021-02-20 #3026 @arfon
@gully @j-faria - would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

The software under review is [_starry_process: Interpretable Gaussian processes for stellar light curves_](https://github.com/rodluger/starry_process).
