### 2021-02-22 6e94c34
arviz hack for v0.11.2 incompat with pymc3
