### 2021-02-22 2488af2
fix arviz issues
