### 2021-03-12 rodluger/starry_process#20 @j-faria
ValueError from Quickstart
Hi @rodluger,

I tried running the first example in the Quickstart and got a `ValueError` when plotting the flux (which boils down to the shape of the output)

```
ValueError: x and y must have same first dimension, but have shapes (1000,) and (1, 1000)
```

(This is part of [joss-reviews/issues/3071](https://github.com/openjournals/joss-reviews/issues/3071))
