### 2021-03-12 rodluger/starry_process#20 @j-faria
In the [docs](https://starry-process.readthedocs.io/en/latest/notebooks/Quickstart/) it's fine.
