### 2021-03-12 rodluger/starry_process#21 @j-faria
[JOSS REVIEW] sp.visualize closes the plot for more than one sample?
When running

```python
y = sp.sample_ylm(nsamples=2).eval()
y.shape # (2, 256)

sp.visualize(y)
```

the plot shows up but closes right after.
There are a few `plt.close()` in `starry_process.visualize.visualize` which are likely the cause.

(This is part of [joss-reviews/issues/3071](https://github.com/openjournals/joss-reviews/issues/3071))
