### 2021-03-12 da83fb1
Fix shape issue in README example
Fixes #20
