### 2021-03-12 rodluger/starry_process#20 @rodluger
Thanks for catching this. The `flux` function returns an array of shape `(nsamples, ntimes)`, so you have to do `flux()[0]` to get the 1d array when only a single sample is requested. I fixed the README -- let me know if that works now.
