### 2021-03-12 rodluger/starry_process#21 @rodluger
Thanks for raising this. It works for me when I do
```bash
python -c "from starry_process import StarryProcess; sp = StarryProcess(); y = sp.sample_ylm(nsamples=2).eval(); sp.visualize(y)"
```
I see a flickering animation switching between the two samples, and the plot remains open until I close it.

Are you in a Jupyter notebook?
