### 2021-03-12 rodluger/starry_process#22 @nespinoza
[JOSS review] List dependencies on README.md and/or documentation
Hey @rodluger!

 A very minor item on my review (https://github.com/openjournals/joss-reviews/issues/3071) is to please list the dependencies of `starry_process` clearly, either on the `README.md` and/or the documentation. I saw these are actually in the `setup.py` (and actually handled automatically as written in the installation instructions!), but listing them might help users figure out what they need (e.g., only Python 3 apparently as a `scipy` version that doesn't support Python 2.7 at least is listed!).

 Thanks in advance!

N.
