### 2021-03-12 rodluger/starry_process#23 @nespinoza
[JOSS review] List Community Guidelines
Hey @rodluger!

 One check we need to do as reviewers of a JOSS submission (https://github.com/openjournals/joss-reviews/issues/3071) are Community Guidelines; basically, a text snippet outlining how to: 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support. While to me the most likely path you are suggesting is to perform these through issue handling/pull requests (as the issue-reporting is implicitly defined in the documentation), I think these should be described explicitly (either in the README and/or documentation) so the workflow is defined without ambiguities for users.

 Thanks again for this fantastic piece of software!
Néstor
