### 2021-03-12 rodluger/starry_process#21 @j-faria
> Are you in a Jupyter notebook?

No, in the IPython console. It's fine if you don't want to support that, I was just surprised with the behaviour.
