### 2021-03-12 #3071 @nespinoza
Hi everyone,

 I have reviewed the submitted manuscript and software presented by @rodluger. This is a fantastic piece of software --- thanks for writing this Rodrigo! I checked all the claims and functionalities following the tutorials in the documentation. I found them very helpful and as a developer that will be implementing this in its own library, I found them to be enlightening on how easy the software is to use. Thanks for that!

 My suggestions, therefore, are all very minor, but are directly related to the checkmarks we must mark as reviewers for JOSS:

 1. I think it would benefit users if you add a section in the `README.md` file on how to install the package (i.e., the `pip install starry-process`). This is explained in the documentation, but folks might check the Github before the docs and wonder how to quickly install the package to try it out (this actually happened to me!).

 2. While the dependencies are listed in the `setup.py` file and are even handled automatically, a clear list of dependencies in text form is still missing. I opened an issue in the repo about this: https://github.com/rodluger/starry_process/issues/22.

 3. I'm also missing some Community Guidelines for contribution. Also opened an issue on this: https://github.com/rodluger/starry_process/issues/23

 Once again, thanks for this software --- if you have any comments regarding this review, please let me know.

Best,
Néstor
