+++ docs/conf.py
@@ -50,7 +50,13 @@ html_css_files = ["custom.css"]
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
 # This pattern also affects html_static_path and html_extra_path.
-exclude_patterns = ["_build", "Thumbs.db", ".DS_Store", "**.ipynb_checkpoints"]
+exclude_patterns = [
+    "_build",
+    "Thumbs.db",
+    ".DS_Store",
+    "**.ipynb_checkpoints",
+    "todo/**",
+]
 
 
 # -- Options for HTML output -------------------------------------------------
