### 2021-03-12 rodluger/starry_process#21 @rodluger
Oh I definitely want to support that! But I still can't reproduce the issue, even inside an IPython session.

What do you get when you run

```python
get_ipython()
```

in your session?
