### 2021-03-12 rodluger/starry_process#21 @j-faria
Ok, I tracked it down to the `--pylab` option to ipython (I know, considered evil!).
Without it, the plot shows and stays flickering.

Regardless of that option, I get this
```python
get_ipython()
<IPython.terminal.interactiveshell.TerminalInteractiveShell at 0x7f9a8f0d2ed0>
```
