### 2021-03-12 rodluger/starry_process#21 @j-faria
Just to add that with `ipython --matplotlib` it also doesn't work.
