### 2021-03-12 rodluger/starry_process#21 @rodluger
Interesting. Seems to be an issue with ``FuncAnimation``. I'll keep digging -- thanks for pointing this out.

This is related:
https://stackoverflow.com/questions/19466825/ipython-pylab-matplotlib-animations-hang
