### 2021-03-12 rodluger/starry_process#21 @j-faria
It works if you return the animation from inside `visualize` and then from `sp.visualize`
