### 2021-03-12 b6e5cca
don't close plot in interactive mode
