### 2021-03-12 rodluger/starry_process#21 @rodluger
Ah, I think I got it. The `--matplotlib` and `--pylab` settings enable interactive mode, in which `plt.show()` is not  blocking, so `plt.close()` immediately closes the animation. (I still want to call `plt.close()` in non-interactive mode, otherwise the plot stays open and calling `visualize` multiple times gobbles up RAM).

The fix is to simply check if `plt.isinteractive()` when calling `plt.close()`. Let me know if the latest commit fixed the issue fore you.
