### 2021-03-15 rodluger/starry_process#22 @rodluger
@nespinoza Thanks for pointing this out. I added a small section to the [README](https://github.com/rodluger/starry_process/blob/master/README.md) stating `starry_process` requires Python 3 as well as a section to the [Installation](https://starry-process.readthedocs.io/en/latest/install/) page on the docs stating the dependencies.
