### 2021-03-15 04ae2dd
aesara compat; misc tweaks
