+++ tests/test_norm.py
@@ -2,7 +2,7 @@ from starry_process.ops import AlphaBetaOp
 from starry_process import StarryProcess
 import numpy as np
 from theano.configparser import change_flags
-import theano.tensor as tt
+from starry_process.compat import theano, tt
 
 
 def test_norm_grad():
