+++ starry_process/compat.py
@@ -1,6 +1,16 @@
 # -*- coding: utf-8 -*-
+from aesara_theano_fallback import aesara as theano
+import aesara_theano_fallback.tensor as tt
+from aesara_theano_fallback import ifelse, USE_AESARA
+from aesara_theano_fallback.tensor import slinalg
+from aesara_theano_fallback.graph import basic, op, params_type, fg
 
 __all__ = [
+    "USE_AESARA",
+    "theano",
+    "tt",
+    "ifelse",
+    "slinalg",
     "Apply",
     "COp",
     "Op",
@@ -12,18 +22,18 @@ __all__ = [
     "random_uniform",
 ]
 
-try:
-    from theano.graph.basic import Apply, Node
-    from theano.graph.op import ExternalCOp as COp
-    from theano.graph.op import Op
-    from theano.graph.params_type import Params, ParamsType
-except ImportError:
-    from theano.gof.graph import Apply, Node
-    from theano.gof.op import COp, Op
-    from theano.gof.params_type import Params, ParamsType
+# Compatibility imports
+Node = basic.Node
+Apply = basic.Apply
+Op = op.Op
+COp = op.ExternalCOp
+Params = params_type.Params
+ParamsType = params_type.ParamsType
+theano.config.floatX = "float64"
 
-try:
-    from theano.tensor.random.utils import RandomStream
+if USE_AESARA:
+
+    from aesara.tensor.random.utils import RandomStream
 
     def random_normal(rng, shape):
         return rng.normal(size=shape)
@@ -32,13 +42,26 @@ try:
         return rng.uniform(size=shape)
 
 
-except ImportError:
-    from theano.tensor.shared_randomstreams import (
-        RandomStreams as RandomStream,
-    )
+else:
 
-    def random_normal(rng, shape):
-        return rng.normal(shape)
+    try:
 
-    def random_uniform(rng, shape):
-        return rng.uniform(shape)
+        from theano.tensor.random.utils import RandomStream
+
+        def random_normal(rng, shape):
+            return rng.normal(size=shape)
+
+        def random_uniform(rng, shape):
+            return rng.uniform(size=shape)
+
+    except ImportError:
+
+        from theano.tensor.shared_randomstreams import (
+            RandomStreams as RandomStream,
+        )
+
+        def random_normal(rng, shape):
+            return rng.normal(shape)
+
+        def random_uniform(rng, shape):
+            return rng.uniform(shape)
