+++ starry_process/ops/flux/L.py
@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
 from ..base_op import BaseOp
-from ...compat import theano, tt
-from ...compat import Apply
+from ...compat import theano, tt, floatX, Apply
 
 __all__ = ["LOp"]
 
@@ -11,13 +10,9 @@ class LOp(BaseOp):
     func_name = "APPLY_SPECIFIC(L)"
 
     def make_node(self, u):
-        in_args = [
-            tt.as_tensor_variable(arg).astype(tt.config.floatX) for arg in [u]
-        ]
+        in_args = [tt.as_tensor_variable(arg).astype(floatX) for arg in [u]]
         out_args = [
-            tt.TensorType(
-                dtype=tt.config.floatX, broadcastable=[False, False]
-            )()
+            tt.TensorType(dtype=floatX, broadcastable=[False, False])()
         ]
         return Apply(self, in_args, out_args)
 
