+++ starry_process/ops/wigner/Rx.py
@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
 from ..base_op import BaseOp
-from ...compat import Apply
-from ...compat import theano, tt
+from ...compat import theano, tt, floatX, Apply
 
 __all__ = ["RxOp"]
 
@@ -12,12 +11,11 @@ class RxOp(BaseOp):
 
     def make_node(self, theta):
         in_args = [
-            tt.as_tensor_variable(arg).astype(tt.config.floatX)
-            for arg in [theta]
+            tt.as_tensor_variable(arg).astype(floatX) for arg in [theta]
         ]
         out_args = [
-            tt.TensorType(dtype=tt.config.floatX, broadcastable=[False])(),
-            tt.TensorType(dtype=tt.config.floatX, broadcastable=[False])(),
+            tt.TensorType(dtype=floatX, broadcastable=[False])(),
+            tt.TensorType(dtype=floatX, broadcastable=[False])(),
         ]
         return Apply(self, in_args, out_args)
 
