+++ tests/test_norm.py
@@ -11,8 +11,8 @@ def test_norm_grad():
         op = AlphaBetaOp(20)
         get_alpha = lambda z: op(z)[0]
         get_beta = lambda z: op(z)[1]
-        tt.verify_grad(get_alpha, [z], n_tests=1, rng=np.random)
-        tt.verify_grad(get_beta, [z], n_tests=1, rng=np.random)
+        theano.gradient.verify_grad(get_alpha, [z], n_tests=1, rng=np.random)
+        theano.gradient.verify_grad(get_beta, [z], n_tests=1, rng=np.random)
 
 
 def test_norm(ftol=0.05):
