+++ starry_process/ops/wigner/special_tensordotRz_rev.py
@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
 from ..base_op import BaseOp
-from ...compat import Apply
-from ...compat import theano, tt
+from ...compat import Apply, theano, tt, floatX
 
 __all__ = ["special_tensordotRzRevOp"]
 
@@ -12,14 +11,12 @@ class special_tensordotRzRevOp(BaseOp):
 
     def make_node(self, T, M, theta, bf):
         in_args = [
-            tt.as_tensor_variable(arg).astype(tt.config.floatX)
+            tt.as_tensor_variable(arg).astype(floatX)
             for arg in [T, M, theta, bf]
         ]
         out_args = [
-            tt.TensorType(
-                dtype=tt.config.floatX, broadcastable=[False, False]
-            )(),
-            tt.TensorType(dtype=tt.config.floatX, broadcastable=[False])(),
+            tt.TensorType(dtype=floatX, broadcastable=[False, False])(),
+            tt.TensorType(dtype=floatX, broadcastable=[False])(),
         ]
         return Apply(self, in_args, out_args)
 
