+++ starry_process/ops/wigner/special_tensordotRz.py
@@ -1,8 +1,7 @@
 # -*- coding: utf-8 -*-
 from ..base_op import BaseOp
-from ...compat import Apply
+from ...compat import Apply, theano, tt, floatX
 from .special_tensordotRz_rev import special_tensordotRzRevOp
-from ...compat import theano, tt
 
 __all__ = ["special_tensordotRzOp"]
 
@@ -17,12 +16,9 @@ class special_tensordotRzOp(BaseOp):
 
     def make_node(self, T, M, theta):
         in_args = [
-            tt.as_tensor_variable(arg).astype(tt.config.floatX)
-            for arg in [T, M, theta]
+            tt.as_tensor_variable(arg).astype(floatX) for arg in [T, M, theta]
         ]
-        out_args = [
-            tt.TensorType(dtype=tt.config.floatX, broadcastable=[False])()
-        ]
+        out_args = [tt.TensorType(dtype=floatX, broadcastable=[False])()]
         return Apply(self, in_args, out_args)
 
     def infer_shape(self, *args):

[…]
