### 2021-03-15 rodluger/starry_process#23 @rodluger
@nespinoza Thanks for raising this. I added a CONTRIBUTING.md file at the root of the repo and now mention it in the README.
