### 2021-03-15 #3071 @j-faria
I finished reviewing the paper and the software. There were a couple of minor issues with the README (some that @nespinoza also identified) and a plot/animation; they were all fixed.
Everything else looks good so all lights are green from my side to accept this submission.

This is an outstanding piece of work, congratulations to @rodluger and the other contributors.
I'm looking forward to using `starry_process` myself so I'll be sure to keep in touch to see what your plans are regarding RVs :)

PS: only the app doesn't work for me, but I'll check if it's my problem before opening an issue. Nothing that should delay publication.
