### 2021-03-19 #3071 @nespinoza
Thanks for all the additions @rodluger. On my end, I can call this review complete then --- happy to accept this paper for publication in JOSS on my end. Congrats @rodluger on this awesome piece of software!
