### 2021-03-30 481839f
always run notebooks
