### 2021-04-06 #3071 @rodluger
@arfon The two associated AAS papers are still in review (one has just been accepted, the other should be accepted in the next couple days). Is there a way to synchronize the release to get the cross-refs correct?

https://ui.adsabs.harvard.edu/abs/2021arXiv210200007L/abstract
https://ui.adsabs.harvard.edu/abs/2021arXiv210201697L/abstract
