### 2021-04-11 #3071 @arfon
@rodluger - yes, the AAS office should be asking me for the DOI of this paper. Do you have the DOIs for the AAS papers yet? We can cite them both in this paper, and link to one of them in the side-bar too using the metadata in the JOSS paper Markdown header (although you will have to pick just one of them for this).
