### 2021-06-22 #3071 @arfon
I just heard from the AAS that their DOI is `10.3847/1538-3881/abfdb9`.

@rodluger - could you add the relevant information to your `paper.md`:

```
aas-doi: 10.3847/1538-3881/abfdb9
aas-journal: Astrophysical Journal <- (or whatever AAS journal you submitted to)
```
