### 2021-07-10 #3071 @arfon
@rodluger – friendly bump on this. Could you please add the AAS paper DOI to your JOSS paper?
