### 2021-07-10 #3071 @whedon
:wave: @openjournals/joss-eics, this paper is ready to be accepted and published.
